refactor(schema): consolidate datetime types into single module
